review
- Use `GIT_CONFIG_PARAMETERS` instead of fiddling with configuration files
