+++ tests/tools/.gitignore
@@ -1,3 +1,4 @@
-fixtures/
+tests/fixtures/*
+!tests/fixtures/*.sh
 !src/signature/fixtures/
 !src/signature/fixtures/**
